refactor(epg): extract epg list helpers
